requirements.txt
sqlalchemy
pandas
numpy
fastapi
python-dotenv
sentence_transformers
faiss-cpu
pyodbc
rapidfuzz